panic if the netlink connection has gone stale
In the current exporter design, the netlink connection is opened once at exporter
startup. If OVS is restarted or upgraded, the exporter will therefore hold a stale
connection. Instead of persistently return something like a NewInvalidMetric, it
is more effective to trigger a panic, and let a process watchdog like systemd
bring the exporter back to life.